add feature if i delete supplier products delete auto

+++ backend/dabactock/src/main/java/dabastock/om/dabactock/services/SupplierService.java
@@ -2,16 +2,23 @@ package dabastock.om.dabactock.services;
 
 import dabastock.om.dabactock.model.Product;
 import dabastock.om.dabactock.model.Supplier;
+import dabastock.om.dabactock.repository.ProductRepository;
 import dabastock.om.dabactock.repository.SupplierRepository;
+import org.hibernate.mapping.Array;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class SupplierService {
     @Autowired
     SupplierRepository supplier_repo;
+
+    @Autowired
+    ProductRepository product_repo;
 
     public List<Supplier> getSuppliers() {
         return supplier_repo.findAll();
@@ -34,6 +41,26 @@ public class SupplierService {
     }
 
     public void deleteSupplier(int id) {
-        supplier_repo.deleteById(id);
+        Supplier supplier = supplier_repo.findById(id).orElse(null);
+        if (supplier != null) {
+            // Search In Product How Have this Supplier name
+            String supplier_name = supplier.getSupplier_name();
+            List<Integer> product_id_matched = new ArrayList<>();
+            List<Product> products = product_repo.findAll();
+
+            // Loop Thought Products
+            for (Product product: products) {
+                if (Objects.equals(product.getSupplier_name(), supplier_name)) product_id_matched.add(product.getProduct_id());
+            }
+
+            // Delete Product That match Supplier Name
+            for (int product_id: product_id_matched) {
+                product_repo.deleteById(product_id);
+            }
+
+            supplier_repo.deleteById(id);
+        }
+
+
     }
 }
